Computing effective amplitude of movement.
Using it to get the effective index of difficulty. Adding infos to the final result.

+++ Assets/Scripts/IndexOfDifficulty.cs
@@ -2,7 +2,7 @@
 {
     public float targetWidth { get; private set; }
     public float targetsDistance { get; private set; }
-    public float indexOfDifficulty { get; private set; }
+    public double indexOfDifficulty { get; private set; }
 
     public IndexOfDifficulty(float targetWidth, float targetsDistance)
     {
